Separate header styled components
